Contributed by tony@chromium.org:
Have gclient not update directories that appear to be handled by git.
This allows us to work on src via git, but allows all the DEPS to be
handled by svn.  One can selectively control which deps are using
git and which are using svn.

This is a simple way to use git for main development while using svn
for all dependencies.