We talked about such a change ages ago and then I got caught up in myriad other tasks. This changes gclient to work relative to the directory of the DEPS file rather than the .gclient (solutions) file. This allows dependencies to declare their relative paths in a more modular way, so that the same DEPS can be used in different solutions without knowledge of their solution level directory layout.
Notes on this patch:
1) - This includes my earlier patch: http://codereview.appspot.com/14092/show. 
I can upload a rebased patch when that one goes in, but I need that patch to run
the tests.

2) This is a breaking change for sure. This would require a round of DEPS
updates in chromium and everyone would need to update their depot tools. Not
sure if there is any precedence on how to do such an upgrade.

Patch contributed by Kelly.

Review: http://codereview.appspot.com/14103/show